Use QLocale::formattedDataSize() if available (Qt 5.10+)
If build against Qt 5.10 or newer, internally QLocale::formattedDataSize()
will be used for values that fit into a qint64. The old implementation
is still there for older Qt versions and for values that do not fit into
a qint64.

The main difference bitween the internal implementation in utils.cpp as
calcFileSize() and QLocale::formattedDataSize() is that QLocale also
translates the unit name.