Fix Engine::insertLibrary() and add a test for it
Inserting a library directly with the new insertLibrary() function
added the library to the list of already loaded libraries
(EnginePrivate::m_libraries), but the Parser uses the return value
of Engine::defaultLibraries() to load the tags and filters of each
library.

https://github.com/cutelyst/cutelee/blob/33b629a6af6168a5eedd1be79b90cb3187b35a57/templates/lib/parser.cpp#L90-L109

Adding the name of the inserted library to the list of default libraries
(EnginePrivate::m_defaultLibraries) fixes this issue.

I think that should be fine as Engine::loadDefaultLibraries() and
Engine::loadLibrary() both check if the library has already been loaded.
We might could change the name of insertLibrary() to
insertDefaultLibrary() to make it more clear.