Make enums comparable
I thought enums were already comparable, but this only applied to the
equality operator, not for less than etc. My unit tests for the comparison
operators were successfully completed, but this was apparently rather
coincidental.

This now adds the missing comparison operators to MetaEnumVariable and
registers them with QMetaType::registerComparators().